move StepTitle
